Change how the ModelExporter is loaded
I don't understand why, but if ModelExporter is loaded prior to HdPlugin, and it loads some OpenGL-relevant class, it triggers a very weird issue with calls to `glGenVertexArrays`, where the generated VAO becomes unusable and causes further issues downstream. Having DeveloperTools be responsible for loading the model exporter on plugin startup ensures that HdPlugin gets to initialize OpenGL properly first.